@@ -14,7 +14,7 @@ import 'constants.dart';
 /// Wraps the firebase auth functionality into a service
 class FirebaseAuthenticationService {
   /// An Instance of Logger that can be used to log out what's happening in the service
-  final Logger log;
+  final Logger? log;
 
   /// The URI to which the authorization redirects. It must include a domain name, and can’t be an IP address or localhost.
   ///
@@ -37,8 +37,8 @@ class FirebaseAuthenticationService {
     this.log,
   });
 
-  String _pendingEmail;
-  AuthCredential _pendingCredential;
+  String? _pendingEmail;
+  AuthCredential? _pendingCredential;
 
   Future<UserCredential> _signInWithCredential(
     AuthCredential credential,
@@ -47,18 +47,18 @@ class FirebaseAuthenticationService {
   }
 
   /// Returns the latest userToken stored in the Firebase Auth lib
-  Future<String> get userToken {
+  Future<String>? get userToken {
     return _firebaseAuth.currentUser?.getIdToken();
   }
 
   /// Returns true when a user has logged in or signed on this device
   bool get hasUser {
-    return _firebaseAuth?.currentUser != null;
+    return _firebaseAuth.currentUser != null;
   }
 
   Future<FirebaseAuthenticationResult> signInWithGoogle() async {
     try {
-      final GoogleSignInAccount googleSignInAccount =
+      final GoogleSignInAccount? googleSignInAccount =
           await _googleSignIn.signIn();
       if (googleSignInAccount == null) {
         log?.i('Process is canceled by the user');
@@ -77,13 +77,13 @@ class FirebaseAuthenticationService {
 
       // Link the pending credential with the existing account
       if (_pendingCredential != null) {
-        await result.user.linkWithCredential(_pendingCredential);
+        await result.user?.linkWithCredential(_pendingCredential!);
         _clearPendingData();
       }
 
       return FirebaseAuthenticationResult(
-        uid: result.user.uid,
-        userToken: result.user.refreshToken,
+        uid: result.user?.uid,
+        userToken: result.user?.refreshToken,
       );
     } on FirebaseAuthException catch (e) {
       log?.e(e);
@@ -146,13 +146,13 @@ class FirebaseAuthenticationService {
 
       // Link the pending credential with the existing account
       if (_pendingCredential != null) {
-        await result.user.linkWithCredential(_pendingCredential);
+        await result.user?.linkWithCredential(_pendingCredential!);
         _clearPendingData();
       }
 
       return FirebaseAuthenticationResult(
-        uid: result.user.uid,
-        userToken: result.user.refreshToken,
+        uid: result.user?.uid,
+        userToken: result.user?.refreshToken,
       );
     } on FirebaseAuthException catch (e) {
       log?.e(e);
@@ -167,52 +167,57 @@ class FirebaseAuthenticationService {
 
   Future<FirebaseAuthenticationResult> signInWithFacebook() async {
     try {
-      AccessToken accessToken = await _facebookLogin.login();
-      log?.v('Facebook Sign In complete. \naccessToken:${accessToken.token}');
+      LoginResult loginResult = await _facebookLogin.login();
+      if (loginResult.status == LoginStatus.success) {}
+      log?.v(
+          'Facebook Sign In complete. \naccessToken:${loginResult.accessToken}');
 
-      final FacebookAuthCredential facebookCredentials =
-          FacebookAuthProvider.credential(accessToken.token);
+      final facebookCredentials =
+          FacebookAuthProvider.credential(loginResult.accessToken?.token ?? '');
 
       final result = await _signInWithCredential(facebookCredentials);
 
       // Link the pending credential with the existing account
       if (_pendingCredential != null) {
-        await result.user.linkWithCredential(_pendingCredential);
+        await result.user?.linkWithCredential(_pendingCredential!);
         _clearPendingData();
       }
 
       return FirebaseAuthenticationResult(
-        uid: result.user.uid,
-        userToken: result.user.refreshToken,
+        uid: result.user?.uid,
+        userToken: result.user?.refreshToken,
       );
     } on FirebaseAuthException catch (e) {
       log?.e(e);
       return await _handleAccountExists(e);
-    } on FacebookAuthException catch (e) {
-      log?.e(e);
-      switch (e.errorCode) {
-        case FacebookAuthErrorCode.OPERATION_IN_PROGRESS:
-          return FirebaseAuthenticationResult.error(
-              errorMessage:
-                  'You have a previous Facebook login operation in progress');
-        case FacebookAuthErrorCode.CANCELLED:
-          return FirebaseAuthenticationResult.error(
-              errorMessage: 'Facebook login has been cancelled by the user');
-        case FacebookAuthErrorCode.FAILED:
-          return FirebaseAuthenticationResult.error(
-              errorMessage: 'Facebook login has failed');
-        default:
-          return FirebaseAuthenticationResult.error(errorMessage: e.toString());
-      }
-    } catch (e) {
+      // ignore: nullable_type_in_catch_clause
+    }
+    //  on FacebookAuthException catch (e) {
+    //   log?.e(e);
+    //   switch (e.errorCode) {
+    //     case FacebookAuthErrorCode.OPERATION_IN_PROGRESS:
+    //       return FirebaseAuthenticationResult.error(
+    //           errorMessage:
+    //               'You have a previous Facebook login operation in progress');
+    //     case FacebookAuthErrorCode.CANCELLED:
+    //       return FirebaseAuthenticationResult.error(
+    //           errorMessage: 'Facebook login has been cancelled by the user');
+    //     case FacebookAuthErrorCode.FAILED:
+    //       return FirebaseAuthenticationResult.error(
+    //           errorMessage: 'Facebook login has failed');
+    //     default:
+    //       return FirebaseAuthenticationResult.error(errorMessage: e.toString());
+    //   }
+    // }
+    catch (e) {
       log?.e(e);
       return FirebaseAuthenticationResult.error(errorMessage: e.toString());
     }
   }
 
   Future<FirebaseAuthenticationResult> loginWithEmail({
-    @required String email,
-    @required String password,
+    required String email,
+    required String password,
   }) async {
     try {
       log?.d('email:$email');
@@ -225,13 +230,13 @@ class FirebaseAuthenticationService {
 
       // Link the pending credential with the existing account
       if (_pendingCredential != null) {
-        await result.user.linkWithCredential(_pendingCredential);
+        await result.user?.linkWithCredential(_pendingCredential!);
         _clearPendingData();
       }
 
       return FirebaseAuthenticationResult(
-        uid: result.user.uid,
-        userToken: result.user.refreshToken,
+        uid: result.user?.uid,
+        userToken: result.user?.refreshToken,
       );
     } on FirebaseAuthException catch (e) {
       log?.e('A firebase exception has occured. $e');
@@ -246,8 +251,10 @@ class FirebaseAuthenticationService {
   }
 
   /// Uses `createUserWithEmailAndPassword` to sign up to the Firebase application
-  Future<FirebaseAuthenticationResult> createAccountWithEmail(
-      {String email, String password}) async {
+  Future<FirebaseAuthenticationResult> createAccountWithEmail({
+    required String email,
+    required String password,
+  }) async {
     try {
       log?.d('email:$email');
       final result = await _firebaseAuth.createUserWithEmailAndPassword(
@@ -259,8 +266,8 @@ class FirebaseAuthenticationService {
           'Create user with email result: ${result.credential} ${result.user}');
 
       return FirebaseAuthenticationResult(
-        uid: result.user.uid,
-        userToken: result.user.refreshToken,
+        uid: result.user?.uid,
+        userToken: result.user?.refreshToken,
       );
     } on FirebaseAuthException catch (e) {
       log?.e('A firebase exception has occured. $e');
@@ -286,7 +293,7 @@ class FirebaseAuthenticationService {
 
     // Fetch a list of what sign-in methods exist for the conflicting user
     List<String> userSignInMethods =
-        await _firebaseAuth.fetchSignInMethodsForEmail(_pendingEmail);
+        await _firebaseAuth.fetchSignInMethodsForEmail(_pendingEmail ?? '');
 
     // If the user has several sign-in methods,
     // the first method in the list will be the "recommended" method to use.
@@ -367,14 +374,14 @@ class FirebaseAuthenticationService {
   Future validatePassword(String password) async {
     try {
       var authCredentials = EmailAuthProvider.credential(
-        email: _firebaseAuth.currentUser.email,
+        email: _firebaseAuth.currentUser?.email ?? '',
         password: password,
       );
 
       var authResult = await _firebaseAuth.currentUser
-          .reauthenticateWithCredential(authCredentials);
+          ?.reauthenticateWithCredential(authCredentials);
 
-      return authResult.user != null;
+      return authResult?.user != null;
     } catch (e) {
       log?.e('Could not validate the user password. $e');
       return FirebaseAuthenticationResult.error(
@@ -384,7 +391,7 @@ class FirebaseAuthenticationService {
 
   /// Update the [password] of the Firebase User
   Future updatePassword(String password) async {
-    await _firebaseAuth.currentUser.updatePassword(password);
+    await _firebaseAuth.currentUser?.updatePassword(password);
   }
 
   /// Generates a cryptographically secure random nonce, to be included in a
@@ -407,13 +414,13 @@ class FirebaseAuthenticationService {
 
 class FirebaseAuthenticationResult {
   /// The users unique id
-  final String uid;
+  final String? uid;
 
   /// The authentication token associated with the user
-  final String userToken;
+  final String? userToken;
 
   /// Contains the error message for the request
-  final String errorMessage;
+  final String? errorMessage;
 
   FirebaseAuthenticationResult({this.uid, this.userToken})
       : errorMessage = null;
@@ -423,7 +430,7 @@ class FirebaseAuthenticationResult {
         userToken = null;
 
   /// Returns true if the response has an error associated with it
-  bool get hasError => errorMessage != null && errorMessage.isNotEmpty;
+  bool get hasError => errorMessage != null && errorMessage!.isNotEmpty;
 }
 
 String getErrorMessageFromFirebaseException(FirebaseAuthException exception) {
